Move h/exceptions.py -> h/views/api/exceptions.py
This file contains exceptions that're only used by the API views.
