+++ h/views/api/auth.py
@@ -12,7 +12,7 @@ from pyramid.view import view_config, view_defaults
 
 from h import models
 from h._compat import urlparse
-from h.exceptions import OAuthTokenError
+from h.views.api.exceptions import OAuthTokenError
 from h.services.oauth_validator import DEFAULT_SCOPES
 from h.util.datetime import utc_iso8601
 from h.views.api.config import api_config
