+++ h/views/api/users.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 from pyramid.httpexceptions import HTTPConflict
 
 from h.auth.util import client_authority
-from h.exceptions import PayloadError
+from h.views.api.exceptions import PayloadError
 from h.presenters import UserJSONPresenter
 from h.schemas.api.user import CreateUserAPISchema, UpdateUserAPISchema
 from h.schemas import ValidationError
